Rename class to a more representative name.
These files have almost nothing to do with ODBC.
